requirements.txt
Django
pillow
requests
django-allauth==65.9.0
psycopg2-binary==2.9.10
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
python-decouple
google-generativeai==0.8.5
upstash-redis==1.4.0
dotenv